MALI: mali_hikey: Modify irq initialization in mali hikey platform logic
Due to the irq domain code landing upstream in 3.19, the logic
for requesting irqs in the hikey platform code no longer works.

This patch tries to sovle the issue by using platform_irq_get
to fetch irq details from the DT entry (which properly sets up
the irq through the GIC), and then re-writes the static tables
so irq code gets the right irq number.

This patch cheats a bit, since it only calls platform_irq_get()
once on the first interrupt value in the DT, but that's because
the mali device shares interrupts.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I0f59858803ecd6a933280af1b91f75aa6b375654