mali: Fix mali timeline use-after-free problem
The mali timeline logic has some configuration dependent
memory ownership rules that don't seem totally baked.

This resulted in use-after-free bugs with DEBUG_SLAB enabled:
[   24.707042] Unable to handle kernel paging request at virtual address 6b6b6b6b6b6b6ba3
[   24.715001] pgd = ffffffc05bc23000
[   24.718418] [6b6b6b6b6b6b6ba3] *pgd=0000000000000000, *pud=0000000000000000
[   24.725432] Internal error: Oops: 96000004 [#1] PREEMPT SMP
[   24.731051] CPU: 6 PID: 2083 Comm: Thread-1 Not tainted 4.4.8-01027-g956a9dd-dirty #953
[   24.739094] Hardware name: HiKey Development Board (DT)
[   24.744364] task: ffffffc074d40300 ti: ffffffc05be04000 task.ti: ffffffc05be04000
[   24.751883] PC is at del_timer+0x18/0x78
[   24.755854] LR is at try_to_grab_pending+0x68/0x194
[   24.760771] pc : [<ffffffc00011c9a4>] lr : [<ffffffc0000d723c>] pstate: 600001c5
[   24.768206] sp : ffffffc05be07c00
[   24.771540] x29: ffffffc05be07c00 x28: ffffffc05be04000
[   24.774800] x27: ffffffc000b89000 x26: ffffffc0010ab000
[   24.774817] x25: ffffffc0000d5974 x24: ffffffc05be07cd8
[   24.774831] x23: ffffffc001027618 x22: 0000000000000001
[   24.774844] x21: 0000000000000001 x20: ffffffc05be07cb8
[   24.774858] x19: 6b6b6b6b6b6b6b9b x18: 0000000000000001
[   24.774872] x17: 0000007fb6184c94 x16: ffffffc0002050d4
[   24.774896] x15: 0000007fa86f2f20 x14: 000000000004eb60
[   24.774909] x13: 00000000ffffffff x12: 0000000000000000
[   24.774922] x11: 0000000000000000 x10: 0000000000000001
[   24.774935] x9 : 0000000000000002 x8 : ffffffc0747f3c00
[   24.774949] x7 : 6b6b6b6b6b6b6b6b x6 : 0000000000000018
[   24.774962] x5 : ffffffc00115fa10 x4 : 0000000000000000
[   24.774975] x3 : ffffffc005f00158 x2 : ffffffc05be07cb8
[   24.774987] x1 : 0000000000000001 x0 : ffffffc0000d723c
<skipping register memory and stack memory dump>
[   24.779621] Call trace:
[   24.779642] [<ffffffc00011c9a4>] del_timer+0x18/0x78
[   24.779655] [<ffffffc0000d723c>] try_to_grab_pending+0x68/0x194
[   24.779665] [<ffffffc0000d8800>] __cancel_work_timer+0x50/0x1b0
[   24.779676] [<ffffffc0000d8aa0>] cancel_delayed_work_sync+0x24/0x30
[   24.779695] [<ffffffc0005e31e8>] _mali_osk_wq_delayed_cancel_work_sync+0x20/0x2c
[   24.779710] [<ffffffc00062487c>] mali_timeline_destroy+0x1f4/0x2f8
[   24.779723] [<ffffffc0006276e0>] mali_timeline_system_destroy+0xe8/0x1d4
[   24.779743] [<ffffffc0005f5de8>] _mali_ukk_close+0x100/0x348
[   24.779753] [<ffffffc0005f6cc8>] mali_release+0x40/0xd4
[   24.779769] [<ffffffc000207b34>] __fput+0x98/0x1b4
[   24.779781] [<ffffffc000207cc8>] ____fput+0x20/0x2c
[   24.779793] [<ffffffc0000dc42c>] task_work_run+0xac/0xc8
[   24.779808] [<ffffffc00008a390>] do_notify_resume+0x7c/0x84
[   24.779819] [<ffffffc000085ba8>] work_pending+0x1c/0x20
[   24.779832] Code: a90153f3 aa0003f3 aa1e03e0 d503201f (f9400660)
[   24.779923] ---[ end trace ac26363ba1209219 ]---

As seen in the backtrace above, in mali_timeline_destroy(), we call
_mali_osk_wq_delayed_cancel_work_sync() and hit the bug. The problem
is that when we call _mali_osk_wq_delayed_cancel_work_sync(), the
timeline structure is already freed.

This is because with CONFIG_SYNC, before we call _cancel_work_sync(),
we call sync_timeline_destroy(), which calls sync_timeline_put(), which
triggers sync_timeline_free() which calls release_obj() on the timeline,
which calls timeline_release(), which frees the structure.

The problem being, the structure is still referenced in the
system->timelines[] array, which then tries to destroy them
again (causing the oops).

This fix tries to address this by making the logic more exclusive.
So if we have CONFIG_SYNC, we handle freeing the system timelines
through the sync_timeline_destroy(), and only in !CONFIG_SYNC do
we free the system timelines via mali_timeline_destroy().
(mali_timeline_destroy is still used on errors during allocation
before we assign the system->timelines[] array - and there we
are sure to cancel the work before we call sync_timeline_release())).

Not sure if this is correct, but it seems to work and avoids
the crash.

Signed-off-by: John Stultz <john.stultz@linaro.org>