i2c: designware: Make sure clock is gated in ->remove()
The current approach relies on CONFIG_PM to be set, as it's expected that
the call to pm_runtime_put_sync() will trigger the ->runtime_suspend()
callback to be invoked.

Even in the case when CONFIG_PM is set, userspace may prevent runtime
suspend via sysfs and pm_runtime_forbid() for the device.

The proper solution to this issue, is to do clock gating without involving
runtime suspend, so let's do this.

Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: John Stultz <john.stultz@linaro.org>