k3dma: add support to dma_min_chan
dma_min_chan as a property for k3dma, it defines minimum channel number
which a non-secure mode driver can use.

One sample usage of this is in Hi3660 SoC. DMA channel 0 is reserved to
lpm3, which is a coprocessor for power management. So as a result, any
request in kernel (which runs on main processor and in non-secure mode)
should start from at least channel 1.

Cc: Guodong Xu <guodong.xu@linaro.org>
Signed-off-by: Li Yu <liyu65@hisilicon.com>