drm: hikey960: Switch to CMA allocated fbdev instead of smmu allocated
This brings the hikey960 4.9 code a bit closer to the upstream
tracking code, which avoids the smmu/iommu allocated fbdev and
instead allocates the framebuffer out of CMA.

One downside is that the CMA framebuffer does not support the
out-of-tree FBIOGET_DMABUF ioctl. Due to how the binary
libGLES_mali.so and gralloc interact, gralloc wants to provide a
dmabuf pointing to the framebuffer created by FBDEV_EMULATION.
With this change, it cannot do so, and thus ends up falling back
to the system heap and losing the double-buffering support,
which causes graphics performance to be slow.

Once we switch to using the drm_hwcomposer, this performance
issue will no longer be a problem, as we will drop
FBDEV_EMULATION and allocate the framebuffer in userland via
gralloc/ion.

Signed-off-by: John Stultz <john.stultz@linaro.org>