usb: dwc2: Fix hibernation between host and device modes.
When core is in hibernation in host mode and a device cable
was connected than driver exited from device hibernation.
However, registers saved for host mode and when exited from
device hibernation register restore would be done for device
register which was wrong because there was no device registers
stored to restore.

In handling status change interrupt checking if
driver is in device mode and if the GPWRDN_BSESSVLD
is set than exiting by gpwrdn disconnect detect flow
with the current gpwrdn register value in device mode.
In case when driver is in host mode exiting hibernation
by gpwrdn disconnect detect flow with the current gpwrdn
register in host mode.
If these actions are not taken, after disconnecting the
device cable core will exit from hibernation. Whereas,
we do need core stay in hibernation in that case.

- Added dwc_handle_gpwrdn_disc_det() function which handles
  gpwrdn disconnect detect flow and exits hibernation
  without restoring the registers.
- Added dwc_handle_gpwrdn_idsts_change() function which
  receives the current mode gpwrdn register and
  depending on host or device mode calls
  dwc_handle_gpwrdn_disc_det() function and exits from
  hibernation.

Change-Id: Ide71a48dc2a7ab7a6113f8e11fb171a3f06ec64e
Signed-off-by: Artur Petrosyan <arturp@synopsys.com>
Signed-off-by: Minas Harutyunyan <hminas@synopsys.com>
Signed-off-by: John Stultz <john.stultz@linaro.org>