drm: kirin: Bump the max_height and max_width values
The max_height and max_width values for the kirin driver have
been set at 2048 since the driver was created.

However, I suspect that value was more of a placeholder then
accurate to the hardware.

Looking up details on the mali450 hardware, the max texture size
is 4096x4096, and I've not come across any meaninful limits to
the kirin display hardware.

Recent changes to the cma-helper code in:
  894a677f4b3e ("drm/cma-helper: Use the generic fbdev emulation")
started enforcing these max resolution values on framebuffers
which previously did not happen, and has resulted in problems
recently with AOSP when surfaceflinger allocates a framebuffer
of 1920x2160 (2*Y of 1080 for page-flipping).

Thus this patch fixes this uncovered problem by bumping the
max_height and max_width values to 4096.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I8768408e8cc0eddf0a6b2dce96c0060338740e09