drm: adv7511: Remove bad mode line from adv7511
In testing w/ drm_hwcomposer, I've found we end up
selecting a mode the montior cannot display. This results
in a black screen rather then things working.

Previously this mode was thought to work, but that might
have only been valid on the hikey board and not the
hikey960 board. So remove this from the whitelist for now.

Signed-off-by: John Stultz <john.stultz@linaro.org>