dts: hikey960: Fixup graphics errors on 4gb boards
On the 4gb boards, the cma region we use for the
framebuffer and ion heap was being placed in memory
mapped above 32bits.

Due to addresses being masked to 32bits for DMA, this
resulted in crashes on newer 4gb boards.

This patch sets the CMA address to be lower in memory
so this doesn't happen and resolves the problem.

I've picked an address just past a previous CMA
reserved chunk, so we shouldn't be adding to any
fragmentation.

Extra thanks to Ryan Grachek <ryan@edited.us> who
noticed what the source of the problem was.

Change-Id: Ibb7ff1d85a9b93e41c440ffacf6a1ccf6aecb1ca
Signed-off-by: John Stultz <john.stultz@linaro.org>