misc: hisi_hikey_usb: Fix USB-C power handling
We've had some behavioral quirks with USB-C power handling as
the board will try to stay on with power removed, drawaing from
the usb-c port. This isn't reliable, so use a fix from the
patches being upstreamed.

This also resolves issue with USB-C plug having to be removed
rather then the vbus power in order to switch to host mode.

Originally-by: Yu Chen <chenyu56@huawei.com>
[jstultz: Ported oneline fix back to android-hikey-linaro-4.19]
Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ie6a3e945297f672ef47b97a1c25814cce43a8866