Fix for Image.gz-dtb support
Commit 06995804b5762 ("arm64: Use full path in KBUILD_IMAGE definition")
unfortunately hard-coded the filename part of the KBUILD_IMAGE for the
default GNU Make target ("all") which means that an Image.gz-dtb will
not be generated by default, even if all of the rules are in place to
make it, *unless* it is specified explicitly by target name. This is not
desirable as we want the Image.gz-dtb file by default whenever this
kbuild feature is enabled.

Bug: 121043321
Change-Id: I6b9857ae5c2cfbeeebdb65fc0e6c816f1b21f4c6
Signed-off-by: Alistair Strachan <astrachan@google.com>