usb: typec: Rationalize the API for the muxes
We can replace the second parameter that is passed to the
typec_mux_get() function with alt mode description
structure, and simply pass NULL with accessory modes.

With accessory modes there is no need for additional
identification (in practice the accessory mode can only be
Audio Accessory if muxing is needed), only with alternate
modes we need to identify the exact alternate mode.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>