usb: typeci_rt1711h: Fix error with devcons not being removed in error path
I was seeing a number of crashes which was due to
device_connection_add() being called on a connector from
rt1711h_probe(), but then when the probe errors out,
the connection is left pointing to the devm alloced
structure which will eventually be freed.

Thus this patch calls device_conneciton_remove() in
the error path to avoid the problem.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: I3189736dcb5c1efb03f841a24a5b9540f03bde3e